requirements.txt
keras
matplotlib==3.3.4
numpy==1.22.1
pandas==1.2.5
tensorflow
Pillow==9.2.0
scipy==1.6.3
